Clean resource to build Instagram application
