+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
-        android:icon="@drawable/cloudy"
+        android:icon="@drawable/ins_app_ic"
         android:label="@string/app_name"
-        android:roundIcon="@drawable/cloudy"
+        android:roundIcon="@drawable/ins_app_ic"
         android:supportsRtl="true"
         android:theme="@style/Theme.Basesrc"
         tools:targetApi="31">
