+++ app/src/main/java/com/base/base_source/activities/MainActivity.kt
@@ -101,7 +101,6 @@ class MainActivity : AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
             R.id.action_userInfo -> {
-                openInformationFragment()
                 return true
             }
 
@@ -109,11 +108,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun openInformationFragment() {
-        findNavController(R.id.nav_host_fragment_content_main).navigate(
-            R.id.action_FirstFragment_to_userInformationFragment
-        )
-    }
+
 
     override fun onSupportNavigateUp(): Boolean {
         val navController = findNavController(R.id.nav_host_fragment_content_main)
