--- app/src/main/java/com/base/base_source/ui/details/DetailVM.kt
@@ -1,8 +1,0 @@
-package com.base.base_source.ui.details
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class DetailVM @Inject constructor() : ViewModel()
