--- app/src/main/java/com/base/base_source/ui/weather/WeatherViewModel.kt
@@ -1,3 +1,0 @@
-package com.base.base_source.ui.weather
-
-class WeatherViewModel
