--- app/src/main/res/drawable/bg_gradien.xml
@@ -1,10 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <gradient
-                android:startColor="#FFEE9B"
-                android:endColor="#FFEE9B"
-                android:angle="90"/>
-        </shape>
-    </item>
-</layer-list>
