--- app/src/main/res/drawable/circle_caret_down.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="666dp"
-    android:height="666dp"
-    android:viewportWidth="666"
-    android:viewportHeight="666">
-  <path
-      android:pathData="M333,333m-333,0a333,333 0,1 1,666 0a333,333 0,1 1,-666 0"
-      android:fillColor="#ffffff"/>
-  <path
-      android:pathData="M166.04,237C156.33,237 147.49,241.33 143.37,248.11C139.24,254.9 140.58,262.9 146.8,268.66L313.76,423.06C318.52,427.46 325.57,430 333,430C340.43,430 347.48,427.46 352.24,423.06L519.2,268.66C525.42,262.9 526.76,254.9 522.64,248.11C518.51,241.33 509.67,237 499.96,237H166.04Z"
-      android:fillColor="#1196FF"/>
-</vector>
