--- app/src/main/res/drawable/circle_caret_up.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="674dp"
-    android:height="673dp"
-    android:viewportWidth="674"
-    android:viewportHeight="673">
-  <path
-      android:pathData="M337,336.38m332.98,-3.36a333,333 0,1 1,-665.97 6.72a333,333 0,1 1,665.97 -6.72"
-      android:fillColor="#ffffff"/>
-  <path
-      android:pathData="M504.47,431.25C514.19,431.15 522.98,426.73 527.04,419.91C531.09,413.09 529.67,405.1 523.39,399.4L354.89,246.7C350.08,242.35 343.01,239.87 335.58,239.95C328.14,240.02 321.12,242.64 316.41,247.09L151.02,403.16C144.86,408.98 143.6,417 147.79,423.74C151.98,430.48 160.86,434.72 170.58,434.62L504.47,431.25Z"
-      android:fillColor="#FF6C30"/>
-</vector>
