--- app/src/main/res/drawable/outline_arrow_drop_down_24_red.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24"
-    android:tint="@android:color/holo_orange_dark">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M7,10l5,5 5,-5H7z"/>
-</vector>
