--- app/src/main/res/layout/item_weather.xml
@@ -1,74 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="160dp"
-    android:layout_height="100dp"
-    android:orientation="vertical"
-    android:layout_margin="8dp">
-
-    <RelativeLayout
-        android:padding="12dp"
-        android:background="@drawable/item_background"
-        android:layout_width="140dp"
-        android:layout_height="80dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        >
-
-        <TextView
-            android:id="@+id/tv_time"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="12:00 am"
-            android:textColor="@android:color/white"
-            android:textSize="16sp" />
-
-        <TextView
-            android:id="@+id/tv_temperature"
-            android:layout_width="wrap_content"
-            android:layout_below="@id/tv_time"
-            android:layout_height="wrap_content"
-            android:gravity="bottom"
-            android:layout_alignParentBottom="true"
-            android:text="47°"
-            android:textColor="@android:color/white"
-            android:textSize="28sp" />
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:layout_alignParentBottom="true"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:layout_width="20dp"
-                android:layout_height="20dp"
-                android:layout_gravity="center"
-                android:layout_marginEnd="10dp"
-                android:src="@drawable/water_drop" />
-
-            <TextView
-                android:id="@+id/tv_precipitation"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="5%"
-                android:textColor="@android:color/white"
-                android:textSize="28sp" />
-        </LinearLayout>
-
-    </RelativeLayout>
-
-
-    <ImageView
-        android:id="@+id/iv_weather_icon"
-        android:layout_width="55dp"
-        android:layout_height="55dp"
-        android:layout_gravity="center"
-        android:src="@drawable/item_img_1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
