--- app/src/main/res/layout/item_weather_vertical.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="70dp"
-        android:layout_marginBottom="6dp"
-        android:background="@drawable/item_background"
-        android:orientation="horizontal"
-        android:paddingHorizontal="12dp"
-        tools:ignore="UselessParent">
-
-        <ImageView
-            android:id="@+id/imageViewz"
-            android:layout_width="60dp"
-            android:layout_height="60dp"
-            android:layout_alignParentStart="true"
-            android:layout_centerVertical="true"
-            android:layout_marginStart="0dp"
-            android:src="@drawable/cloudy" />
-
-        <TextView
-            android:id="@+id/textView5"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentEnd="true"
-            android:layout_centerVertical="true"
-            android:text="57º"
-            android:textSize="16sp"
-            android:textColor="@color/white" />
-
-        <ImageView
-            android:id="@+id/circleCaretUp"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:layout_centerVertical="true"
-            android:layout_marginEnd="8dp"
-            android:layout_toStartOf="@+id/textView5"
-            android:src="@drawable/circle_caret_up" />
-
-        <TextView
-            android:id="@+id/textView7"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerVertical="true"
-            android:layout_marginEnd="16dp"
-            android:textSize="16sp"
-            android:layout_toStartOf="@+id/circleCaretUp"
-            android:text="47º"
-            android:textColor="@color/white" />
-
-        <ImageView
-            android:id="@+id/circleCaretDown"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:layout_centerVertical="true"
-            android:layout_marginEnd="8dp"
-            android:layout_toStartOf="@+id/textView7"
-            android:src="@drawable/circle_caret_down" />
-
-        <TextView
-            android:id="@+id/textViewContent"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerVertical="true"
-            android:layout_marginStart="8dp"
-            android:layout_marginEnd="8dp"
-            android:textSize="16sp"
-            android:layout_toStartOf="@+id/circleCaretDown"
-            android:layout_toEndOf="@+id/imageViewz"
-            android:text="Cloudy"
-            android:textColor="@color/white" />
-
-    </RelativeLayout>
-
-</LinearLayout>
