+++ app/src/main/res/navigation/nav_graph.xml
@@ -2,36 +2,10 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/calculatorFragment">
+    app:startDestination="@id/homeFragment">
 
     <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.base.base_source.ui.syncdata.SyncFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_map">
-        <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
-        <action
-            android:id="@+id/action_FirstFragment_to_userInformationFragment"
-            app:destination="@id/userInformationFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.base.base_source.ui.details.DetailFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
-
-        <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/userInformationFragment"
-        android:name="com.base.base_source.ui.userInformation.UserInformationFragment"
-        android:label="UserInformationFragment"></fragment>
-    <fragment
-        android:id="@+id/calculatorFragment"
-        android:name="com.base.base_source.ui.weather.WeatherFragment"
-        android:label="CalculatorFragment" />
+        android:id="@+id/homeFragment"
+        android:name="com.base.base_source.ui.home.HomeFragment"
+        android:label="HomeFragment" />
 </navigation>
